+++ JsonToolsNppPlugin/Properties/AssemblyInfo.cs
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("5.8.0.11")]
-[assembly: AssemblyFileVersion("5.8.0.11")]
+[assembly: AssemblyVersion("5.8.0.12")]
+[assembly: AssemblyFileVersion("5.8.0.12")]